// part9/patientor-frontend/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Diagnosis['code'][];
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: {
    startDate: string;
    endDate: string;
  }
}

export interface HospitalEntry extends BaseEntry {
  type: "Hospital";
  discharge: {
    date: string;
    criteria: string;
  }
}

export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;

//type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;

//export type EntryWithoutId = UnionOmit<Entry, 'id'>;
export type EntryWithoutId = Omit<HealthCheckEntry | HospitalEntry | OccupationalHealthcareEntry, 'id'>;

export type EntryFormState = 
  | { type: 'HealthCheck'; healthCheckRating: HealthCheckRating; description: string; date: string; specialist: string; diagnosisCodes?: string[]; }
  | { type: 'Hospital'; discharge: { date: string; criteria: string; }; description: string; date: string; specialist: string; diagnosisCodes?: string[]; }
  | { type: 'OccupationalHealthcare'; employerName: string; sickLeave?: { startDate: string; endDate: string; }; description: string; date: string; specialist: string; diagnosisCodes?: string[]; };

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: Gender;
  ssn?: string;
  dateOfBirth?: string;
  entries: Entry[];
}

export type PatientFormValues = Omit<Patient, "id" | "entries">;